requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
sentence-transformers
faiss-cpu
python-dotenv
passlib[bcrypt]
python-jose
redis
